Rough Final Draft
This final copy includes a css sheet and a javascript sheet. I also made sure to import javascript and include jquery.  The biggest change from any other versions is that this file includes the google books API. I also added a  function that populates an HTML divison with information about the book that was searched for.